Refactor main menu buttons into separate widget
