+++ lib/main.dart
@@ -1,4 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:treasure_of_the_high_seas/screens/app.dart';
+import 'screens/app.dart';
 
 void main() => runApp(MyApp());
