+++ lib/screens/app.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:treasure_of_the_high_seas/screens/main_menu.dart';
+import './main_menu/main_menu.dart';
 
 class MyApp extends StatelessWidget {
   // This widget is the root of your application.
